fix: cmake stuff

+++ build.sh
@@ -4,8 +4,10 @@
 BUILD_TYPE=Debug
 Fortran_COMPILER=gfortran
 GENERATOR="Unix Makefiles"
-ENABLE_OPI_SUPPORT=OFF
+ENABLE_OPI_SUPPORT=ON
+ENABLE_PFUNIT=ON
 LIBSUFFIX="so"
+GENERATOR_FLAGS=""
 if [[ "$OSTYPE" == "linux-gnu" ]]; then
   LIBSUFFIX="so"
 elif [[ "$OSTYPE" == "darwin"* ]]; then
@@ -16,7 +18,7 @@ elif [[ "$OSTYPE" == "MINGW"* ]]; then
   LIBSUFFIX="dll"
 elif [[ "$OSTYPE" == "msys"* ]]; then
   LIBSUFFIX="dll"
-  GENERATOR="MSYS Makefiles"
+  GENERATOR_FLAGS="-G MSYS Makefiles"
 fi
 git submodule update --init --recursive
 ################################################################################
@@ -26,16 +28,16 @@ git submodule update --init --recursive
 ################################################################################
 if [[ "$ENABLE_OPI_SUPPORT" == "ON" ]]; then
   # Build OPI
-  cd OPI || exit || exit
+  cd OPI || exit
   # Create the build directory if it does not exist
   if [[ ! -d "build" ]]; then
     mkdir build
   else
     rm -rf build/*
   fi
-  cd build || exit || exit
+  cd build || exit
   echo "Updating cmake"
-  cmake -DCMAKE_BUILD_TYPE=$BUILD_TYPE -DCMAKE_INSTALL_PREFIX=../ -DCMAKE_Fortran_COMPILER=$Fortran_COMPILER -DENABLE_CXX11=ON -DENABLE_CL_SUPPORT=OFF -DENABLE_CUDA_SUPPORT=OFF -DENABLE_PYTHON=OFF -G "$GENERATOR" ../
+  cmake -DCMAKE_BUILD_TYPE=$BUILD_TYPE -DCMAKE_INSTALL_PREFIX=../ -DCMAKE_Fortran_COMPILER=$Fortran_COMPILER -DENABLE_CXX11=ON -DENABLE_CL_SUPPORT=OFF -DENABLE_CUDA_SUPPORT=OFF -DENABLE_PYTHON=OFF "$GENERATOR_FLAGS" ../
   echo "Building OPI"
   cmake --build .
   cmake --install .
@@ -45,6 +47,8 @@ if [[ "$ENABLE_OPI_SUPPORT" == "ON" ]]; then
   fi
   echo "Leaving OPI"
   cd ../../
+else
+  echo "ENABLE_OPI_SUPPORT is 'OFF'. Skipping. Set to 'ON' to add it."
 fi
 ################################################################################
 #                                                                              #
@@ -61,7 +65,7 @@ else
 fi
 cd build || exit
 echo "Updating cmake"
-cmake -DCMAKE_BUILD_TYPE=$BUILD_TYPE -DCMAKE_Fortran_COMPILER=$Fortran_COMPILER -DENABLE_OpenMP_SUPPORT=OFF -DENABLE_POSTGRESQL_SUPPORT=OFF -G "$GENERATOR" ../
+cmake -DCMAKE_BUILD_TYPE=$BUILD_TYPE -DCMAKE_Fortran_COMPILER=$Fortran_COMPILER -DENABLE_OpenMP_SUPPORT=OFF -DENABLE_POSTGRESQL_SUPPORT=OFF "$GENERATOR_FLAGS" ../
 echo "Building libslam"
 cmake --build .
 cmake --install .
@@ -89,23 +93,28 @@ cd ../
 #                                Build pFUnit                                  #
 #                                                                              #
 ################################################################################
-cd pFUnit || exit
-# Create the build directory if it does not exist
-if [[ ! -d "build" ]]; then
-  mkdir build
+if [[ "$ENABLE_PFUNIT" == "ON" ]]; then
+  cd pFUnit || exit
+  # Create the build directory if it does not exist
+  if [[ ! -d "build" ]]; then
+    mkdir build
+  else
+    rm -rf build/*
+  fi
+  cd build || exit
+  echo "Updating cmake"
+  export PFUNIT_DIR=..//pFUnit/build/installed
+  export FC=$Fortran_COMPILER
+  cmake -DSKIP_MPI=yes "$GENERATOR_FLAGS" ../
+  echo "Building pFUnit"
+  cmake --build .
+  cmake --install .
+  cd ../
+  echo "Leaving pFUnit"
+  cd ../
 else
-  rm -rf build/*
+  echo "ENABLE_PFUNIT is 'OFF'. Skipping. Set to 'ON' to add it."
 fi
-cd build || exit
-echo "Updating cmake"
-export FC=$Fortran_COMPILER
-cmake -DSKIP_MPI=yes -G "$GENERATOR" ../
-echo "Building pFUnit"
-cmake --build .
-cmake --install .
-cd ../
-echo "Leaving pFUnit"
-cd ../
 ################################################################################
 #                                                                              #
 #                                 Build NEPTUNE                                #
@@ -116,7 +125,7 @@ if [[ ! -d "lib" ]]; then
   mkdir lib
 fi
 # Create the links to libraries needed by NEPTUNE
-cd lib || exit || exit
+cd lib || exit
 ln -sf ../libslam/lib/libslam-Fortran.$LIBSUFFIX .
 if [[ "$ENABLE_OPI_SUPPORT" == "ON" ]]; then
   ln -sf ../OPI/lib/libOPI-Fortran.$LIBSUFFIX .
@@ -142,8 +151,7 @@ else
 fi
 cd build || exit
 echo "Updating cmake"
-export PFUNIT_DIR=..//pFUnit/build/installed
-cmake -DCMAKE_BUILD_TYPE=$BUILD_TYPE -DCMAKE_Fortran_COMPILER=$Fortran_COMPILER -DENABLE_OPI_SUPPORT=$ENABLE_OPI_SUPPORT -DSKIP_MSIS_2=ON -G "$GENERATOR" ../
+cmake -DCMAKE_BUILD_TYPE=$BUILD_TYPE -DCMAKE_Fortran_COMPILER=$Fortran_COMPILER -DENABLE_OPI_SUPPORT=$ENABLE_OPI_SUPPORT -DSKIP_MSIS_2=ON -DENABLE_PFUNIT=$ENABLE_PFUNIT "$GENERATOR_FLAGS" ../
 echo "Building NEPTUNE"
 cmake --build .
 cmake --install .
@@ -152,15 +160,15 @@ if [[ $? -ne 0 ]]; then
     exit $?
 fi
 echo "Leaving NEPTUNE"
-cd ../work || exit || exit
+cd ../work || exit
 if [[ "$OSTYPE" == "msys"* || "$OSTYPE" == "MINGW"* ]]; then
-  ln -sf ../bin/neptune-sa.exe . || exit || exit
-  ln -sf ../bin/libneptune.dll . || exit || exit
-  ln -sf ../lib/libneptune.dll.a . || exit || exit
-  ln -sf ../libslam/lib/libslam-Fortran.dll . || exit || exit
-  ln -sf ../libslam/lib/libslam-Fortran.dll.a . || exit || exit
+  ln -sf ../bin/neptune-sa.exe . || exit
+  ln -sf ../bin/libneptune.dll . || exit
+  ln -sf ../lib/libneptune.dll.a . || exit
+  ln -sf ../libslam/lib/libslam-Fortran.dll . || exit
+  ln -sf ../libslam/lib/libslam-Fortran.dll.a . || exit
 else 
-  ln -sf ../bin/neptune-sa . || exit || exit
+  ln -sf ../bin/neptune-sa . || exit
 fi
 cd ..
 echo "Done"
